+++ CSharp/MotionController/MotionController.Sensor.Db.Data/Models/GameSessionInputData.cs
@@ -17,10 +17,10 @@ public sealed class GameSessionInputData : DatabaseModel
     public int GameSessionId { get; set; }
 
     [Column("X")]
-    public decimal X { get; set; }
+    public float X { get; set; }
 
     [Column("Y")]
-    public decimal Y { get; set; }
+    public float Y { get; set; }
 
     [Column("Created")]
     public DateTime Created { get; set; }
